+++ nacc_attribute_deriver/attribute_deriver.py
@@ -203,7 +203,8 @@ class MissingnessDeriver(BaseAttributeDeriver):
                 rule_schema = MissingnessFileModel.model_validate(row)
 
                 # generally consider 1 to 1 mapping so throw error on duplicates
-                if rule_schema.function in results:
+                # except for headers which are defined multiple times
+                if rule_schema.function in results and not rule_schema.function.startswith("header_"):
                     raise AttributeDeriverError(
                         f"Multiple missingness rules defined for {rule_schema.function}"
                     )