+++ nacc_attribute_deriver/attributes/missingness/modules/ftld/missingness_ftld.py
@@ -34,11 +34,3 @@ class FTLDFormMissingness(FormMissingnessCollection):
             return 9999
 
         return self.generic_missingness("ftdinfyr", int)
-
-    def _missingness_ftld_formver(self) -> float:
-        """Handles missingness for FTLD FORMVER."""
-        return self.generic_missingness("formver", float)
-
-    def _missingness_ftld_adcid(self) -> int:
-        """Handles missingness for FTLD ADCID."""
-        return self.generic_missingness("adcid", int)