+++ nacc_attribute_deriver/attributes/missingness/modules/lbd/missingness_lbd.py
@@ -21,11 +21,3 @@ class LBDFormMissingness(FormMissingnessCollection):
     def _missingness_lbd(self, field: str, attr_type: Type[T]) -> T:
         """Defines general missingness for LBD form variables."""
         return self.generic_missingness(field, attr_type)
-
-    def _missingness_lbd_formver(self) -> float:
-        """Handles missingness for LBD FORMVER."""
-        return self.generic_missingness("formver", float)
-
-    def _missingness_lbd_adcid(self) -> int:
-        """Handles missingness for LBD ADCID."""
-        return self.generic_missingness("adcid", int)